// client/state/Products/products.action-types.ts
export enum ActionTypes {
  FETCH_PRODUCTS_START = 'FETCH_PRODUCTS_START',
  FETCH_PRODUCTS_SUCCESS = 'FETCH_PRODUCTS_SUCCESS',
  FETCH_PRODUCTS_ERROR = 'FETCH_PRODUCTS_ERROR',

  FETCH_PRODUCT_START = 'FETCH_PRODUCT_START',
  FETCH_PRODUCT_SUCCESS = 'FETCH_PRODUCT_SUCCESS',
  FETCH_PRODUCT_ERROR = 'FETCH_PRODUCT_ERROR',
  FETCH_PRODUCT_RESET = 'FETCH_PRODUCT_RESET',

  DELETE_PRODUCT_START = 'DELETE_PRODUCT_START',
  DELETE_PRODUCT_SUCCESS = 'DELETE_PRODUCT_SUCCESS',
  DELETE_PRODUCT_ERROR = 'DELETE_PRODUCT_ERROR',
}